+++ Sources/KeyboardKit/Extensions/String+Chars.swift
@@ -12,4 +12,9 @@ public extension String {
     
     /// Split the string into a list of characters.
     var chars: [String] { map(String.init) }
+    
+    /// Split the string into a list of character actions.
+    var charActions: [KeyboardAction] {
+        chars.map(KeyboardAction.character)
+    }
 }